+++ express-joi-validation.js
@@ -59,17 +59,8 @@ function buildErrorString(err, container) {
   return ret
 }
 
-module.exports = function() {
-  throw new Error(
-    'express-joi-validation: exported member is no longer a factory function. use exported createValidator function instead'
-  )
-}
-
 module.exports.createValidator = function generateJoiMiddlewareInstance(cfg) {
   cfg = cfg || {} // default to an empty config
-
-  const Joi = cfg.joi || require('@hapi/joi')
-
   // We'll return this instance of the middleware
   const instance = {
     response
@@ -81,9 +72,9 @@ module.exports.createValidator = function generateJoiMiddlewareInstance(cfg) {
 
     instance[type] = function(schema, opts) {
       opts = opts || {} // like config, default to empty object
-
-      return function exporessJoiValidator(req, res, next) {
-        const ret = Joi.validate(req[type], schema, opts.joi || container.joi)
+      const computedOpts = { ...container.joi, ...cfg.joi, ...opts.joi }
+      return function expressJoiValidator(req, res, next) {
+        const ret = schema.validate(req[type], computedOpts)
 
         if (!ret.error) {
           req[container.storageProperty] = req[type]
@@ -111,7 +102,7 @@ module.exports.createValidator = function generateJoiMiddlewareInstance(cfg) {
       next()
 
       function validateJson(json) {
-        const ret = Joi.validate(json, schema, opts.joi)
+        const ret = schema.validate(json, opts.joi)
         const { error, value } = ret
         if (!error) {
           // return res.json ret to retain express compatibility